Update fr.ts
I have looked through and fixed some typos of the French translation. 

I just have to note for the line 11, day, when we count in French, most numbers will finish in "me": seventy fifth is soixante-quinzième etc. The only difference is the first day, which would end in "r" (premier jour). Maybe it would be possible to add a separate line for the first day or this could simply be disregarded.